feat(graph): centralize shared render policy
