+++ packages/graph-engine/src/render/community-washes.ts
@@ -1,4 +1,4 @@
-import type { RenderableCommunity } from "./model";
+import type { RenderableCommunity } from "./render-policy";
 
 const SVG_NS = "http://www.w3.org/2000/svg";
