+++ packages/graph-engine/src/render/edges.ts
@@ -1,4 +1,4 @@
-import type { RenderableEdge } from "./model";
+import type { RenderableEdge } from "./render-policy";
 
 const SVG_NS = "http://www.w3.org/2000/svg";
