+++ packages/graph-engine/src/render/sigma-global-renderer.ts
@@ -10,7 +10,7 @@ import {
 import type { GraphScreenPoint } from "./geometry";
 import type { GraphGestureTarget } from "./gestures";
 import type { GraphRendererAdapterData } from "./adapter";
-import { edgeRelationClass } from "./model";
+import { edgeRelationClass } from "./render-policy";
 import { DEFAULT_RENDERER_VIEWPORT, type RendererViewport, type RendererViewportSize } from "./viewport";
 import {
   emptyGraphFirstOrderRelationFocusTouched,
