+++ packages/graph-engine/test/fixtures/issue-159/dependency-baseline.json
@@ -26,13 +26,5 @@
     }
   ],
   "internalModelBarrelReferences": [],
-  "rendererRouteBypasses": [
-    {
-      "source": "render/sigma-global-renderer.ts",
-      "target": "render/model.ts",
-      "specifier": "./model",
-      "kind": "import",
-      "typeOnly": false
-    }
-  ]
+  "rendererRouteBypasses": []
 }
